backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
sqlalchemy==2.0.30
pydantic==2.7.1
pydantic-settings==2.3.4
python-dotenv==1.0.1
langchain==0.2.5
langchain-google-genai==1.0.6
langchain-community==0.2.5
chromadb==0.4.24
faiss-cpu==1.8.0
google-generativeai==0.5.4
pypdf==4.2.0
tiktoken==0.7.0
httpx==0.27.0
aiofiles==23.2.1